+++ src/Logic/Sellmachine.java
@@ -3,20 +3,20 @@ package Logic;
 public class Sellmachine {
 
 
-public void holzverkauf(double holzmenge,int percent) {
+public void holzverkauf(long holzmenge,int percent) {
 if (holzmenge >=0) {
 	Gui.Gui.banc.deposit(holzmenge*50);
 }
 
 }
 
-public void steinverkauf(double steinmenge,int percent) {
+public void steinverkauf(long steinmenge,int percent) {
 	if (steinmenge >=0) {
 		Gui.Gui.banc.deposit(steinmenge*80);
 	}
 }
 
-public void goldverkauf(double goldmenge,int percent) {
+public void goldverkauf(long goldmenge,int percent) {
 	if (goldmenge >=0) {
 		Gui.Gui.banc.deposit(goldmenge*200);
 	}
